turn player_ids into floats for consistency in case of nulls

+++ player_scraper.py
@@ -4,7 +4,7 @@ def flatten_player(player, game_id):
     player_data = player[1]
 
     return {'game_id': game_id,
-            'player_id': int(player_id),
+            'player_id': float(player_id),
             'team_id': player_data['teamId'],
             'position_id': player_data['posnId'],
             'substitute': player_data['substitute'],
@@ -21,6 +21,6 @@ def flatten_sub(sub, game_id):
             'sub_type': sub_data['type'],
             'time_of_event(min)': (sub_data['t']['m'] + (sub_data['t']['s'] / 60 )),
             'team_id': sub_data['team'],
-            'player_off': sub_data['offId'],
-            'player_on': sub_data['inId']
+            'player_off': float(sub_data['offId']),
+            'player_on': float(sub_data['inId'])
             }
